daily_report_aggregator/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
yfinance>=0.2.30
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
lxml>=4.9.0
pymupdf>=1.23.0
playwright>=1.40.0
